Refactoring: created /types and /constants, added TODOs
