+++ client/package.json
@@ -13,6 +13,7 @@
     "@reduxjs/toolkit": "^2.2.3",
     "@types/chart.js": "^2.9.41",
     "antd": "^5.16.4",
+    "axios": "^1.7.9",
     "chart.js": "^4.4.2",
     "chartjs-adapter-date-fns": "^3.0.0",
     "classnames": "^2.5.1",
