+++ client/src/components/FleetBar/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Item from './Item';
-import { FleetItemType } from '../GoogleMap';
+import { FleetItemType } from '@/types/fleet.type';
 
 function FleetBar({ fleets, activeFleet, disabled, setActiveFleet }: { fleets: FleetItemType[]; activeFleet: string | number | null; disabled: boolean; setActiveFleet: React.Dispatch<React.SetStateAction<string | number | null>> }) {
     return (
