+++ client/src/types/coordinate.type.ts
@@ -1,0 +1,4 @@
+export interface CoordinatesType {
+  lat: number;
+  lng: number;
+}
